fixed-bugs

+++ utils/fileDelete.js
@@ -1,9 +1,11 @@
 const fs = require("fs");
 const fileDelete = (path) => {
-  fs.unlink(path, (err) => {
-    if (err) throw new Error(err.message);
-    return console.log("file was deleted");
-  });
+  if (fs.existsSync(psth)) {
+    fs.unlink(path, (err) => {
+      if (err) throw new Error(err.message);
+      return console.log("file was deleted");
+    });
+  }
 };
 
 module.exports = fileDelete;
